fix rerendering issue
